Adding diagnostics of incorrect closure initialisation expression count -- too little expressions are diagnosed as an error, and too many expressions are correctly diagnosed and extra expressions are dropepd.
Currently, initialisation of closure variables drops values differently
from variabe initialisation -- for closure variable initialisation, the
last (rightmost) expressions are dropped, and for simple variables the
first expressions (the leftmost ones) are dropped. The previously
encoded behaviour of simple variables seems incorrect; it will be
changed to match the behavior of the closre variable initialisation.

git-svn-id: svn+ssh://saulius.grazulis.lt/home/saulius/svn-repositories/compilers/trunk@2359 f7c8cbe2-f9e6-0310-9050-d7f47bf15973